+++ codebase/components/Home/ButtonWrapper.jsx
@@ -1,0 +1,7 @@
+import { styled } from '@mui/system';
+
+const ButtonWrapper = styled('div')(({ theme }) => ({
+  marginTop: theme.spacing(2),
+}));
+
+export default ButtonWrapper;
