+++ codebase/components/Home/ContentContainer.js
@@ -1,0 +1,11 @@
+import { styled } from '@mui/system';
+
+const ContentContainer = styled('div')({
+  display: 'flex',
+  flexDirection: 'column',
+  justifyContent: 'center',
+  alignItems: 'center',
+  maxWidth: '100%',
+});
+
+export default ContentContainer;
