+++ codebase/components/Home/Image.js
@@ -1,0 +1,9 @@
+import { styled } from '@mui/system';
+
+const Image = styled('img')(({ theme }) => ({
+  width: '100%',
+  maxWidth: '500px',
+  margin: theme.spacing(2, 0),
+}));
+
+export default Image;
