+++ codebase/components/Home/Index.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { AppBar, Toolbar, Container, ThemeProvider } from '@mui/material';
 import Header from '../Layout/Header/Index';
 import Footer from '../Layout/Footer/Index';
-import HeroComponent from './Hero';
+import HeroComponent from './HeroComponent';
 import theme from '../styles/theme';
 
 const HomePage = () => {
